Narrow up for use in filter of grid view
If separator empty then hide separator and assume wanting to use this
as flag for a few minor changes. Tested with spin, date time, date, and
money widgets. Money widget needs a merge of PR 110 to allowNull for
placeholders to show if allowZero is true. Spin works best with
verticalButtons. DateTime works best with widget type of TYPE_INPUT.

In order to apply, separator should be empty.

Not really clean for the GridView css to be in the field-range.css
file, but similarly couples GridView to FieldRange if done the other
way. No doubt there is a tidier way to achieve this. The css is only
written for how I am am using it.